Bug 1294260 - Part 3: fix some compile warnings in mailnews that look syntactic only (hints and help from aleth and jorgk). r=jorgk, r=aleth CLOSED TREE
--HG--
extra : amend_source : 657132180b3913c71d4626a41b221ad95656ea98

+++ mailnews/local/src/nsMsgLocalStoreUtils.cpp
@@ -302,7 +302,7 @@ nsMsgLocalStoreUtils::DiskSpaceAvailableInStore(nsIFile *aFile, uint64_t aSpaceR
   nsresult rv = aFile->GetDiskSpaceAvailable(&diskFree);
   if (NS_SUCCEEDED(rv)) {
 #ifdef DEBUG
-    printf("GetDiskSpaceAvailable returned: %lld bytes\n", diskFree);
+    printf("GetDiskSpaceAvailable returned: %lld bytes\n", (long long)diskFree);
 #endif
     // When checking for disk space available, take into consideration
     // possible database changes, therefore ask for a little more
